@@ -1,28 +1,22 @@
 package com.gymicompany.gestionescolar;
-
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Context;
-import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.content.Intent;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
-
 import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.FirebaseFirestore;
-import com.google.firebase.firestore.FirebaseFirestore;
-import com.google.firebase.firestore.Query;
-import com.google.firebase.firestore.QueryDocumentSnapshot;
-import com.google.firebase.firestore.QuerySnapshot;
+import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.firebase.firestore.DocumentSnapshot;
+
 
 
 public class IniciarSesionActivity extends AppCompatActivity {
@@ -37,6 +31,7 @@ public class IniciarSesionActivity extends AppCompatActivity {
     EditText contraseñaeditText, correoeditText;
     FirebaseFirestore mFirestore;
     FirebaseAuth mAuth;
+    FirebaseUser user;
 
     //filtro de usuarios docentes
     FirebaseFirestore db = FirebaseFirestore.getInstance();
@@ -56,6 +51,7 @@ public class IniciarSesionActivity extends AppCompatActivity {
         mFirestore = FirebaseFirestore.getInstance();
         mAuth = FirebaseAuth.getInstance();
         buttonvolver = findViewById(R.id.button_atras);
+        user=FirebaseAuth.getInstance().getCurrentUser();
 
         buttonIniciarSesion.setOnClickListener(new View.OnClickListener() {
 
@@ -91,27 +87,44 @@ public class IniciarSesionActivity extends AppCompatActivity {
             @Override
             public void onComplete(@NonNull Task<AuthResult> task) {
                 if (task.isSuccessful()){
-                    finish();
-                    startActivity(new Intent(IniciarSesionActivity.this, HubAlumnos.class));
-                    Toast.makeText(IniciarSesionActivity.this, "Bienvenido", Toast.LENGTH_SHORT).show();
-                }else{
+                    redirigirUsuario();
+                }
+                else{
                     Toast.makeText(IniciarSesionActivity.this, "Error", Toast.LENGTH_SHORT).show();
                 }
             }
-        }).addOnFailureListener(new OnFailureListener(){
-            @Override
-            public void onFailure(@NonNull Exception e){
-                Toast.makeText(IniciarSesionActivity.this, "error de sesion", Toast.LENGTH_SHORT).show();
-            }
-        });
+        }).addOnFailureListener(e -> Toast.makeText(IniciarSesionActivity.this, "error de sesion", Toast.LENGTH_SHORT).show());
 
+
+
+
+        }
+
+    private void redirigirUsuario() {
+        String userId = FirebaseAuth.getInstance().getCurrentUser().getUid();
+
+        FirebaseFirestore.getInstance().collection("user")
+                .document(userId)
+                .get()
+                .addOnSuccessListener(new OnSuccessListener<DocumentSnapshot>() {
+                    @Override
+                    public void onSuccess(DocumentSnapshot documentSnapshot) {
+                        if (documentSnapshot.exists()) {
+                            String tipoUsuario = documentSnapshot.getString("perfil");
+
+                            // Redirige al usuario según el valor del campo
+                            if ("alumno".equals(tipoUsuario)) {
+                                // Redirige a la pantalla de administrador
+                                Toast.makeText(IniciarSesionActivity.this, "Bienvenido Alumno", Toast.LENGTH_SHORT).show();
+                                startActivity(new Intent(IniciarSesionActivity.this, HubAlumnos.class));
+                            } else {
+                                // Redirige a la pantalla de usuario normal
+                                Toast.makeText(IniciarSesionActivity.this, "Bienvenido profesor", Toast.LENGTH_SHORT).show();
+                                startActivity(new Intent(IniciarSesionActivity.this, hub_docentes.class));
+                            }
+                        }
+                    }
+                });
     }
-
-
-
-
-
-
-
 
 }
